requirements.txt
pyyaml
openai
groq
mecab-python3
unidic-lite
